Refactor everything to new SQL schema
To reduce storage size, schema altered so that all FKs
use the corresponding serial id of the dominant entity
instead of extensive attributes (blockhash, txid etc.).
Also moves the [app|db|rpc]Config.json outside the src/ since
they are provided by the docker-compose.yml of the
satamotodb project and they don't depend on the build
of the typescript.